# apps/be/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API 문서",
    "title": "My API",
    "contact": {},
    "version": "1.0"
  },
  "host": "Localhost:4000",
  "basePath": "/api",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "아이디와 비밀번호로 로그인하여 Access/Refresh Token 발급",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "로그인",
        "parameters": [
          {
            "description": "로그인 요청",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.LoginRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.LoginResponseDTO"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Refresh Token을 무효화하고 쿠키에서 제거합니다.",
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "로그아웃",
        "responses": {
          "200": {
            "description": "로그아웃 성공",
            "schema": {
              "$ref": "#/definitions/dto.LogoutResponseDTO"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Refresh Token을 사용하여 새로운 Access Token을 발급합니다.",
        "produces": ["application/json"],
        "tags": ["auth"],
        "summary": "Access Token 갱신",
        "responses": {
          "200": {
            "description": "accessToken 응답",
            "schema": {
              "$ref": "#/definitions/dto.RefreshTokenResponseDTO"
            }
          },
          "401": {
            "description": "Refresh Token 누락 또는 유효하지 않음",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "description": "storeID에 해당하는 카테고리 이름 목록을 조회합니다.",
        "produces": ["application/json"],
        "tags": ["category"],
        "summary": "카테고리 목록 조회",
        "responses": {
          "200": {
            "description": "카테고리 리스트",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.GetCategoryResponseDTO"
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "조회 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "storeID에 해당하는 카테고리를 추가합니다.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["category"],
        "summary": "카테고리 생성",
        "parameters": [
          {
            "description": "카테고리 요청 데이터",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateCategoryRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "생성된 카테고리 응답",
            "schema": {
              "$ref": "#/definitions/dto.CreateCategoryResponseDTO"
            }
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/categories/{categoryId}": {
      "delete": {
        "description": "카테고리 ID와 storeID로 카테고리를 삭제합니다.",
        "produces": ["application/json"],
        "tags": ["category"],
        "summary": "카테고리 삭제",
        "parameters": [
          {
            "type": "integer",
            "description": "카테고리 ID",
            "name": "categoryId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "삭제 성공 메시지",
            "schema": {
              "$ref": "#/definitions/dto.DeleteCategoryResponseDTO"
            }
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "404": {
            "description": "카테고리 없음",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "description": "카테고리 ID와 storeID로 카테고리 이름을 수정합니다.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["category"],
        "summary": "카테고리 수정",
        "parameters": [
          {
            "type": "integer",
            "description": "카테고리 ID",
            "name": "categoryId",
            "in": "path",
            "required": true
          },
          {
            "description": "수정할 카테고리 정보",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.UpdateCategoryRequestDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.UpdateCategoryResponseDTO"
            }
          },
          "400": {
            "description": "잘못된 요청",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "404": {
            "description": "카테고리 없음",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/menus": {
      "get": {
        "description": "storeID에 해당하는 메뉴 목록을 조회합니다.",
        "produces": ["application/json"],
        "tags": ["menu"],
        "summary": "메뉴 목록 조회",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.GetMenuResponseDTO"
              }
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/menus/": {
      "post": {
        "description": "여러 메뉴에 대해 생성/수정/삭제 동기화를 처리합니다.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["menu"],
        "summary": "메뉴 동기화",
        "parameters": [
          {
            "description": "동기화 요청 데이터",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.SyncMenuRequestDTO"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "전체 성공",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "207": {
            "description": "일부 실패",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "400": {
            "description": "요청 바디 오류",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/menus/board/{categoryID}": {
      "get": {
        "description": "카테고리 ID로 메뉴, 이미지, 태그를 함께 조회합니다.",
        "produces": ["application/json"],
        "tags": ["menu"],
        "summary": "메뉴판 조회",
        "parameters": [
          {
            "type": "integer",
            "description": "카테고리 ID",
            "name": "category_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dto.GetMenuBoardResponseDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    },
    "/menus/{menuID}": {
      "get": {
        "description": "storeID와 menuID로 메뉴 상세 정보를 조회합니다. (Preview, Details, Tags, Images 포함)",
        "produces": ["application/json"],
        "tags": ["menu"],
        "summary": "메뉴 상세 조회",
        "parameters": [
          {
            "type": "integer",
            "description": "메뉴 ID",
            "name": "menuId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/dto.GetDescriptionResponseDTO"
            }
          },
          "400": {
            "description": "잘못된 요청 (menuId 없음 또는 invalid)",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "401": {
            "description": "인증 실패",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          },
          "500": {
            "description": "서버 에러",
            "schema": {
              "$ref": "#/definitions/dto.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.CreateCategoryRequestDTO": {
      "type": "object",
      "required": ["category"],
      "properties": {
        "category": {
          "type": "string",
          "example": "증류주"
        }
      }
    },
    "dto.CreateCategoryResponseDTO": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "음료"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "storeId": {
          "type": "integer",
          "example": 2
        }
      }
    },
    "dto.DeleteCategoryResponseDTO": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "dto.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "unauthorized"
        }
      }
    },
    "dto.GetCategoryResponseDTO": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "음료"
        },
        "id": {
          "type": "integer",
          "example": 1
        }
      }
    },
    "dto.GetDescriptionResponseDTO": {
      "type": "object",
      "properties": {
        "details": {
          "type": "string"
        },
        "images": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "preview": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "dto.GetMenuBoardResponseDTO": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string"
        },
        "category_id": {
          "type": "integer"
        },
        "details": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "image_urls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "menu": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "preview": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "store_id": {
          "type": "integer"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "dto.GetMenuResponseDTO": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string"
        },
        "categoryId": {
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "menu": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "storeId": {
          "type": "integer"
        }
      }
    },
    "dto.LoginRequestDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "example": "1234"
        },
        "userId": {
          "type": "string",
          "example": "admin123"
        }
      }
    },
    "dto.LoginResponseDTO": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR..."
        }
      }
    },
    "dto.LogoutResponseDTO": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "logout success"
        }
      }
    },
    "dto.RefreshTokenResponseDTO": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
    "dto.SyncMenuRequestDTO": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "additionalProperties": {}
        },
        "id": {
          "type": "integer"
        }
      }
    },
    "dto.UpdateCategoryRequestDTO": {
      "type": "object",
      "required": ["category"],
      "properties": {
        "category": {
          "type": "string",
          "example": "증류주"
        }
      }
    },
    "dto.UpdateCategoryResponseDTO": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "음료"
        },
        "id": {
          "type": "integer",
          "example": 1
        },
        "storeId": {
          "type": "integer",
          "example": 2
        }
      }
    }
  }
}